moved to pages-compress files
