+++ criu/Makefile.crtools
@@ -11,7 +11,6 @@ obj-y			+= bitmap.o
 obj-y			+= cgroup.o
 obj-y			+= cgroup-props.o
 obj-y			+= clone-noasan.o
-obj-y			+= compression.o
 obj-y			+= cr-check.o
 obj-y			+= cr-dedup.o
 obj-y			+= cr-dump.o
