--- criu/compression.h
@@ -1,6 +1,0 @@
-#ifndef __CR_COMPRESSION_H__
-#define __CR_COMPRESSION_H__
-
-int decompress_image(int comp_fd, const char *path);
-
-#endif
